VR Course Final Project
Along with a team, I developed a VR game where users can interact with each other using musical instruments. In addition to interaction with other users, the game also allowed a teacher to interact with the users to help them learn music. The languages used for this game was C# and JavaScript. I also integrated the HTC-Vive with its controllers to allow users to play the drums.